Move common rescoring utils to util package
